services/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["pavelhardzei <pavelgordei11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.37"
alembic = "^1.8.0"
psycopg2-binary = "^2.9.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
email-validator = "^1.2.1"
fastapi-mail = "^1.0.9"
alembic-autogen-check = "^1.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
pytest = "^7.1.2"
pytest-factoryboy = "^2.5.0"
freezegun = "^1.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
